requirements.txt
beautifulsoup4
camelot-py
fastapi
geopy
httpx
httpie
metaphone
numpy
pandas
Pillow
PyGithub
python-dotenv
PyYAML
pymupdf4llm
rapidfuzz
requests
uvicorn
yt_dlp
python-multipart
uv
lxml